feat/JUS-89: Finished user registration page
- Finished UI for user registration

- Other frontend refactoring and code cleanup

- Refactored some files/folders and deleted unused css resources
