+++ client/package.json
@@ -14,6 +14,7 @@
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.1.0",
     "react-edit-text": "^5.0.2",
+    "react-google-button": "^0.7.2",
     "react-icons": "^4.4.0",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
