+++ client/public/index.html
@@ -3,7 +3,6 @@
   <head>
     <meta charset="utf-8" />
     <link rel="icon" href="%PUBLIC_URL%/public/favicon.ico" />
-    <link rel="stylesheet"  type="text/css" href="%PUBLIC_URL%/public/stylesheet.css">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
