--- client/public/stylesheet.css
@@ -1,4 +1,0 @@
-html *
-{
-    font-family: "Droid Sans Mono Dotted" !important;
-}
