+++ client/src/components/forms/input/inputField.css
@@ -1,0 +1,12 @@
+@import url('https://fonts.googleapis.com/css?family=Varela+Round&display=swap');
+#input-field {
+    font-size: 1em;
+}
+.icon {
+    margin-top: 10px;
+    margin-left: 12px;
+}
+.form-control:focus {
+    border-color: #ff0000 !important;
+    box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075), 0 0 8px rgba(255, 0, 0, 0.6) !important;
+}
