+++ client/src/components/forms/login/loginform.css
@@ -1,0 +1,20 @@
+#submit-button {
+    box-shadow: 0px 6px 20px -5px #ff0000 !important;
+}
+
+#logo-img {
+    text-align: center;
+}
+
+#login-container {
+    background: #ffffff;
+    color: #465362;
+    height: 700px;
+    width: 700px;
+    border-radius: 20px;
+}
+
+#welcome-text {
+    font-size: 1.5em;
+    font-stretch: expanded;
+}
