+++ client/src/components/forms/signup/signupform.css
@@ -1,0 +1,49 @@
+#boundless-text {
+    color: white;
+    font-size: 2em;
+    margin-top: 45%;
+    background-color: rgb(255, 161, 161);
+    background-image: linear-gradient(-75deg,
+        transparent 0,
+        transparent 5%,
+    rgba(255, 255, 255, 0.5) 5%,
+        rgba(255,255,255,0.5) 10%,
+        transparent 10%,
+        transparent 100%
+    );
+    background-size: 500% 100%;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    transition: 2.5s;
+}
+#boundless-text:hover {
+    background-position: -120% 0%;
+}
+#clicks-text {
+    color: #d7d7d7;
+    font-size: 1.75em;
+}
+#signup-text {
+    font-size: 1.75em;
+}
+#signup-container {
+    background: #ffffff;
+    color: #465362;
+    height: 650px;
+    width: 1100px;
+    border-radius: 20px;
+}
+
+#right-container {
+    border-top-right-radius: inherit;
+    border-bottom-right-radius: inherit;
+    background-image: url('../../../bg.jpg');
+    background-position: top;
+    background-size: cover;
+}
+
+
+@keyframes shine {
+    from { background-position: 0%; }
+    to { background-position: -150%; }
+}
