--- client/src/components/login/loginform.css
@@ -1,49 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Varela+Round&display=swap');
-
-a {
-    color: inherit !important; /* blue colors for links too */
-    text-decoration: inherit !important; /* no underline */
-    cursor: pointer;
-}
-a:focus {
-    outline: 0 !important;
-}
-* {
-    font-family: "Varela Round";
-}
-.form-control:focus {
-     border-color: #ff0000 !important;
-     box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075), 0 0 8px rgba(255, 0, 0, 0.6) !important;
- }
-
-#submit-button {
-    box-shadow: 0px 6px 20px -5px #ff0000 !important;
-}
-
-#logo-img {
-    text-align: center;
-}
-
-#login-container {
-    background: #ffffff;
-    color: #465362;
-    height: 700px;
-    width: 700px;
-    border-radius: 20px;
-}
-
-#welcome-text {
-    font-size: 1.5em;
-    font-stretch: expanded;
-}
-/* style icon */
-.inner-addon .fa {
-    position: absolute;
-    pointer-events: none;
-}
-
-.icon {
-    margin-top: 14px;
-    margin-left: 12px;
-}
-
