+++ client/src/index.js
@@ -2,19 +2,19 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import reportWebVitals from './reportWebVitals'
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
-import Login from './pages/user/login/login.jsx'
-import Signup from './pages/user/signup/signup.jsx'
-import OrganizerRegister from './pages/organizer/register.jsx'
-import OrgVerifyEmail from './pages/organizer/verifyEmail/verifyEmail.jsx'
-import OrgVerificationRequired from './pages/organizer/verificationRequired/verificationRequired.jsx'
-import CustomerVerifyEmail from './pages/customer/verifyEmail/verifyEmail.jsx'
-import CustomerVerificationRequired from './pages/customer/verificationRequired/verificationRequired.jsx'
-import CustomerInfo from './pages/customer/customerInfo/customerInfo.jsx'
+import Login from './pages/user/login.jsx'
+import Signup from "./pages/user/signup.jsx";
+import OrgVerifyEmail from './pages/user/organizer/verifyEmail/verifyEmail.jsx'
+import OrgVerificationRequired from './pages/user/organizer/verificationRequired/verificationRequired.jsx'
+import CustomerVerifyEmail from './pages/user/verifyEmail/verifyEmail.jsx'
+import CustomerVerificationRequired from './pages/user/verificationRequired/verificationRequired.jsx'
+import CustomerInfo from './pages/user/customer/customerInfo/customerInfo.jsx'
 import Header from './components/header/Header'
 import { store } from './store.js'
 import {Provider} from 'react-redux'
 
 import 'bootstrap/dist/css/bootstrap.css'
+import './stylesheet.css'
 
 export default function App() {
   return (
@@ -26,10 +26,9 @@ export default function App() {
             <Route path="personal-info" element={<CustomerInfo />} />
             <Route path="login" element={<Login />} />
             <Route path="signup" element={<Signup />} />
-            <Route path="register" element={<OrganizerRegister />} />
             <Route path="organizer">
               <Route path="verification-required" element={<OrgVerificationRequired />} />
-              <Route path="verifyemail">              
+              <Route path="verifyemail">
                 <Route path=":email">
                   <Route path=":token" element={<OrgVerifyEmail />} />
                 </Route>
