+++ client/src/stylesheet.css
@@ -1,0 +1,13 @@
+@import url('https://fonts.googleapis.com/css?family=Varela+Round&display=swap');
+a {
+    color: inherit !important; /* blue colors for links too */
+    text-decoration: inherit !important; /* no underline */
+    cursor: pointer;
+}
+a:focus {
+    outline: 0 !important;
+}
+*
+{
+    font-family: "Varela Round" !important;
+}
